move logic to reshard.cpp
