+++ ttnn/cpp/ttnn/operations/core/to_memory_config/to_memory_config_op.hpp
@@ -9,6 +9,7 @@
 #include "ttnn/core.hpp"
 #include "ttnn/tensor/tensor.hpp"
 #include "ttnn/run_operation.hpp"
+#include "ttnn/operations/data_movement/sharded/reshard/reshard.hpp"
 #include "ttnn/operations/data_movement/sharded/reshard/device/reshard_op.hpp"
 #include "ttnn/operations/data_movement/sharded/interleaved_to_sharded/device/interleaved_to_sharded_op.hpp"
 #include "ttnn/operations/data_movement/sharded/sharded_to_interleaved/device/sharded_to_interleaved_op.hpp"
@@ -46,14 +47,7 @@ struct ToMemoryConfig {
                         throw std::runtime_error(
                             "dtype cannot be specified when converting sharded tensor to sharded tensor");
                     }
-                    return tt::tt_metal::operation::run(
-                               data_movement::ReshardDeviceOperation{
-                                   .output_mem_config = memory_config,
-                               },
-                               {tensor},
-                               {},
-                               {std::nullopt})
-                        .at(0);
+                    return ttnn::reshard(ttnn::DefaultQueueId, tensor, memory_config, std::nullopt);
                 } else {
                     // for row-major tensors where shard-spec[1] is different for input shard and output shard
